requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
apptools==5.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
configobj==5.0.6
cycler==0.11.0
dill==0.3.4
envisage==6.0.1
fire==0.4.0
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.46.0
h5py==3.6.0
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
imutils==0.5.4
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
llvmlite==0.38.0
Markdown==3.3.7
matplotlib==3.5.2
mayavi==4.7.4
mediapipe==0.8.9.1
msgpack==1.0.3
msgpack-numpy==0.4.7.1
multidict==6.0.2
networkx==2.8
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
poseviz @ file:///home/liamch/poseviz
progressbar2==4.0.0
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyface==7.4.1
Pygments==2.12.0
pyparsing==3.0.8
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-utils==3.1.0
PyWavelets==1.3.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
slidingwindow==0.0.14
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.7.0
tensorpack @ git+https://github.com/ppwwyyxx/tensorpack.git@f8fa2108c7ba85a69e0d7b81f2737d8284fdaa67
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tf-slim @ git+https://github.com/adrianc-a/tf-slim.git@80265a15482b4f81f162a344f13659a855cc5543
threadpoolctl==3.1.0
tifffile==2022.5.4
tqdm==4.64.0
traits==6.3.2
traitsui==7.3.1
transforms3d==0.3.1
typing_extensions==4.2.0
urllib3==1.26.9
vtk==9.1.0
Werkzeug==2.1.2
wrapt==1.14.1
wslink==1.6.5
yarl==1.7.2
zipp==3.8.0
